remote parsing failing

+++ VT_MSPM0G3507_A/src/Core/system.hpp
@@ -314,8 +314,8 @@ namespace System {
         };
         static_assert(sizeof(CAN_ID_J1939) == sizeof(uint32_t));
 
-        uint8_t DLC2Len(DL_MCAN_RxBufElement const *);
-        uint8_t DLC2Len(DL_MCAN_TxBufElement const *);
+        uint32_t DLC2Len(DL_MCAN_RxBufElement const *);
+        uint32_t DLC2Len(DL_MCAN_TxBufElement const *);
         uint32_t len2DLC(uint32_t size);
     }
